#794 : Adjusted SyncInputFormatterTask to handle HL7 messages.
Currently, there are some edgecases, which only apply to HL7, that need to be handled in the SyncInputFormatterTask.
This violates the purpose of SyncInputFormatterTask, which should be able to process the offsets of the input buffer without worrying about a special type of data format.
However, as long as the CSV data and HL7 share the same type of OffsetIterator, this is not possible without heavy adjustments on the OffsetIterator class itself, which should be kept as simple as possible.