Compiled: simplified projection passes multi-threads
In a prior attempt, there was a problem when the thread that executed the compiled pipeline was not the same that traced it.
Passing the executionCtx as a pointer into the 'indexTuplesProxy' lead to (at least potentially) invalid values during execution of the compiled pipeline.
By passing concrete nautilus values to the proxy function, we can avoid that problem.
Using the arena required to make the arenaRef of the arena public, because no prior member was a nautilus value.